requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
joblib==1.4.2
numpy==2.1.1
scikit-learn==1.5.2
scipy==1.14.1
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
